ascii_to_bin write doubles

+++ binary_convert/ascii_to_bin.cpp
@@ -2,11 +2,11 @@
 
 int main(int argc,char**argv)
 {
-  float vec[3];
+  double vec[3];
   while (std::cin>>vec[0]>>vec[1]>>vec[2])
   {
     //std::cout<<vec[0]<<" "<<vec[1]<<" "<<vec[2]<<std::endl;
-    std::cout.write(reinterpret_cast<const char*>(vec),3*sizeof(float));
+    std::cout.write(reinterpret_cast<const char*>(vec),3*sizeof(double));
   }
 
   return 0;
